backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9

# DocStruct pipeline dependencies (from plan)
pdfplumber>=0.10.0
pymupdf>=1.23.0
python-docx>=1.0.0
beautifulsoup4>=4.12.0
markdown>=3.5.0
ebooklib>=0.18.0
python-pptx>=0.6.0
Pillow>=10.0.0
scikit-learn>=1.3.0

